Update EU_2018-790.md please
Hi Eoghan, I'd like to propose a file change to incorporate the changes I've made to your git hub doc. Wjat I did: altered the 'Common Principles', moving parts from other areas of the 'Alignment document'; added the publications section, rationalising it – while adhering to the agreed policy – removing any reference to the author depositing publications and ensuring consistency in the use of the term 'suitable repository'.  Altered the research data section, as agreed during our meeting.  Removed the 'Preservation' section and added preservation to the 'Infrastructure' heading (again as agreed). Did no more with Infrastructure as I know you are working on it. I put the 'Skills' and 'rewards' parts from the Alignment doc under the relevant headings and turned the final part (on multi-stakeholder dialogue) into something we can commit to at the national level. It's 6.25 pages on a Word doc at the moment – not too bad.